+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased changes. Release notes have not yet been written.
 1.2.5
 =====
 * Addition of predicates to filter on header field (with regex), minimum sequence length, and minimum quality score (supports Phred+33 and Phred+64)
+* No longer need to specify -j for JSON pattern files
 * Updates to README.md
 
 1.2.4